+++ test/common/validation.go
@@ -20,7 +20,7 @@ import (
 	"time"
 
 	corev1 "k8s.io/api/core/v1"
-	apiextv1b1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
+	apiextv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"sigs.k8s.io/kubefed/pkg/apis/core/common"
@@ -115,7 +115,7 @@ func ValidKubeFedConfig() *v1beta1.KubeFedConfig {
 			Name:      util.KubeFedConfigName,
 		},
 		Spec: v1beta1.KubeFedConfigSpec{
-			Scope: apiextv1b1.ClusterScoped,
+			Scope: apiextv1.ClusterScoped,
 		},
 	}